moved guild state into guilds.py
